requirements.txt
# ── Core ────────────────────────────────────────────────────────
numpy>=1.24
opencv-python>=4.8
Pillow>=10.0

# ── Deep Learning ───────────────────────────────────────────────
torch>=2.1
torchvision>=0.16
facenet-pytorch>=2.5        # MTCNN + InceptionResnet (FaceNet)
insightface>=0.7            # RetinaFace + ArcFace backbone
onnxruntime>=1.16

# ── Liveness / CV ───────────────────────────────────────────────
scikit-learn>=1.3           # SVM for texture analysis
scikit-image>=0.21          # LBP feature extraction
albumentations>=1.3 
mediapipe==0.10.9      # Augmentation pipelines
imbalanced-learn>=0.12
tqdm
uniface
# ── API ─────────────────────────────────────────────────────────
fastapi>=0.110
uvicorn[standard]>=0.29
python-multipart>=0.0.9
pydantic>=2.5

# ── Config & Utilities ──────────────────────────────────────────
PyYAML>=6.0
python-dotenv>=1.0

# ── Testing ─────────────────────────────────────────────────────
pytest>=8.0
httpx>=0.27                 # FastAPI TestClient dependency
